Do not compute debug names of reflection objects (#86880)
We're spending 5% of Stage2 app startup computing debug names.

This code originates from .NET Native where we had a "release mode with developer experience mode enabled" mode that tried to make troubleshooting trimming issues easier. One of the things it did was providing names of things in the debugger. Since we lack compiler support for this, developer experience mode is always enabled right now.

Since in .NET 5+ people are not supposed to troubleshoot trimming issues in the debugger, this mode doesn't make much sense (we also don't have a managed debugger to surface the `DebuggerDisplay` anyway)