Fix IsFixupPrecodeByASM and StubPrecode::IsStubPrecodeByASM (#89899)
These functions use memcmp that can internally be optimized
by vectorization. It can thus read more bytes than necessary
for the purpose of this function. In edge cases it can end up
reading from unmapped memory and crashing with access violation.

The fix is to compare the data byte by byte and early out as soon
as a mismatch is found.