Add test for using fallback RID when reading an empty RID graph (#91040)
When there is no RID graph and the host is configured to use the RID graph, it will discard the computed machine RID (because it is not in the RID graph) and use the fallback RID (compile-time define) to look for assets. This adds a test that exact matches to the fallback RID should be resolved in that case.